move the exit modal to the layout page
