+++ components/studentPortal/layout/Layout.tsx
@@ -1,15 +1,23 @@
-import { useQuery } from "@apollo/client";
-import { addDays, format } from "date-fns";
+import { useMutation, useQuery } from "@apollo/client";
+import { addDays, differenceInHours, format } from "date-fns";
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import {
-  FetchGoalCountResponse,
+  FETCH_LAST_SEEN_MODAL,
+  FetchModalData,
+} from "../../../graphql/studentPortal/freemium/fetchLastSeenModal";
+import { UPSERT_LAST_SEEN_MODAL } from "../../../graphql/studentPortal/freemium/upsertLastSeenModal";
+import {
   FETCH_USER_GOALS_COUNT,
+  FetchGoalCountResponse,
 } from "../../../graphql/studentPortal/goals/fetchUserGoalsCount";
 import { useAuth } from "../../../lib/authContext";
+import { calculateRemainingTrialDays } from "../../../pages/api/studentPortal/freemium/helpers";
 import { profileSelector } from "../../../redux/profileSlice";
 import { setIsGoalApproaching } from "../../../redux/sidebarSlice";
-import { setTheme, Theme, themeSelector } from "../../../redux/themeSlice";
+import { Theme, setTheme, themeSelector } from "../../../redux/themeSlice";
+import FreemiumDialogComponent from "../freemium/FreemiumDialogueComponent";
+import FreemiumExitComponent from "../freemium/FreemiumExitComponent";
 import { FreemiumHeader } from "../freemium/FreemiumHeader";
 import Sidebar from "./Sidebar";
 
@@ -34,6 +42,41 @@ export const Layout: React.FC = ({ children }) => {
       }
     },
     fetchPolicy: "cache-and-network",
+  });
+
+  const [updateLastSeenModal] = useMutation(UPSERT_LAST_SEEN_MODAL);
+  const [showOnboardingModal, setShowOnboardingModal] = useState(false);
+  const [showExitModal, setShowExitModal] = useState(false);
+
+  useQuery<FetchModalData>(FETCH_LAST_SEEN_MODAL, {
+    variables: {
+      userId: user.uid,
+    },
+    skip: userRole != "paid" && userRole != "freemium",
+
+    onCompleted: (data) => {
+      const trialDaysRemaining = calculateRemainingTrialDays(createdAt);
+
+      const lastSeenValue = data.freemium_users[0]?.lastSeenModal;
+
+      const lastSeenDifference = lastSeenValue
+        ? differenceInHours(new Date(), new Date(lastSeenValue))
+        : null;
+
+      // check whether trial has expired and show exit modal
+      if (trialDaysRemaining === 0) {
+        console.log("createdAt", createdAt);
+        setShowExitModal(true);
+      } else {
+        // check whether user has seen onboarding modal in the last 24 hours
+        if (lastSeenDifference > 24 || !lastSeenValue) {
+          setShowOnboardingModal(true);
+          updateLastSeenModal({
+            variables: { userId: user.uid, lastSeenModal: new Date() },
+          });
+        }
+      }
+    },
   });
 
   return (
@@ -98,6 +141,8 @@ export const Layout: React.FC = ({ children }) => {
       >
         <Sidebar />
       </div>
+      {showOnboardingModal && <FreemiumDialogComponent />}
+      {showExitModal && <FreemiumExitComponent />}
     </div>
   );
 };
